Move GroupSlicer to a separate file, add operator+ and +=
