+++ Framework/Core/test/test_GroupSlicer.cxx
@@ -108,7 +108,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerOneAssociated)
   BOOST_CHECK_EQUAL(t.size(), 10 * 20);
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
   for (auto& slice : g) {
@@ -179,7 +179,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerSeveralAssociated)
   BOOST_CHECK_EQUAL(tu.size(), 10 * 20);
 
   auto tt = std::make_tuple(tx, ty, tz, tu);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
   for (auto& slice : g) {
@@ -238,7 +238,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerMismatchedGroups)
   BOOST_CHECK_EQUAL(t.size(), 10 * (20 - 5));
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
   for (auto& slice : g) {
@@ -294,7 +294,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerMismatchedUnassignedGroups)
   BOOST_CHECK_EQUAL(t.size(), (30 + 10 * (20 - 5)));
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
   for (auto& slice : g) {
@@ -342,7 +342,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerMismatchedFilteredGroups)
   BOOST_CHECK_EQUAL(t.size(), 10 * (20 - 4));
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
 
@@ -395,7 +395,7 @@ BOOST_AUTO_TEST_CASE(GroupSlicerMismatchedUnsortedFilteredGroups)
   BOOST_CHECK_EQUAL(t.size(), 10 * (20 - 4));
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
 
@@ -435,7 +435,7 @@ BOOST_AUTO_TEST_CASE(EmptySliceables)
   BOOST_CHECK_EQUAL(t.size(), 0);
 
   auto tt = std::make_tuple(t);
-  o2::framework::AnalysisDataProcessorBuilder::GroupSlicer g(e, tt);
+  o2::framework::GroupSlicer g(e, tt);
 
   unsigned int count = 0;
   for (auto& slice : g) {
